fix errors

+++ src/main/java/com/IntranetManagement/IntranetManagement/Services/DepartmentService.java
@@ -25,19 +25,16 @@ public class DepartmentService {
     }
 
     public List<Department> getAllDepartments() {
+
         return departmentRepository.findAll();
     }
 
-    public Department getDepartmentById(Long departmentId) {
+    public Department getDepartmentById(Integer departmentId) {
         return departmentRepository.findById(departmentId)
                 .orElseThrow(() -> new RuntimeException("Department not found"));
     }
 
-
-
-
-
-    public Department updateDepartment(Long departmentId, String newName, Long userId) {
+    public Department updateDepartment(Integer departmentId, String newName, Long userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RuntimeException("User not found"));
 
@@ -53,7 +50,7 @@ public class DepartmentService {
     }
 
 
-    public void deleteDepartment(Long departmentId, Long userId) {
+    public void deleteDepartment(Integer departmentId, Long userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RuntimeException("User not found"));
 
@@ -71,7 +68,7 @@ public class DepartmentService {
 
 
 
-    public void assignUserToDepartment(Long userId, Long departmentId, Long adminId) {
+    public void assignUserToDepartment(Long userId, Integer departmentId, Long adminId) {
         User admin = userRepository.findById(adminId)
                 .orElseThrow(() -> new RuntimeException("Admin not found"));
 
@@ -96,7 +93,7 @@ public class DepartmentService {
 
 
 
-    public void removeUserFromDepartment(Long userId, Long departmentId, Long adminId) {
+    public void removeUserFromDepartment(Long userId, Integer departmentId, Long adminId) {
         User admin = userRepository.findById(adminId)
                 .orElseThrow(() -> new RuntimeException("Admin not found"));
